% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_names.R
\name{draw_a_name}
\alias{draw_a_name}
\title{Plots the popularity of names in times using the_name and the_sex as arguments}
\usage{
draw_a_name(the_name, the_sex)
}
\arguments{
\item{the_sex}{}
}
\value{
a graph
}
\description{
Plots the popularity of names in times using the_name and the_sex as arguments
}
\examples{
\dontrun{
draw_a_name(the_name,the_sex)
}
}
