% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_names.R
\name{draw_names}
\alias{draw_names}
\title{Simultaneously plot the popularity of more than one name}
\usage{
draw_names(names)
}
\arguments{
\item{names}{vector of several names for which we want to plot the popularity}
}
\value{
a graph
}
\description{
Simultaneously plot the popularity of more than one name
}
\examples{
\dontrun{
draw_names(vector of names I want to plot)
}
}
