% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_excel.R
\name{multi_excel}
\alias{multi_excel}
\title{Import all sheets from an xlsx file}
\usage{
multi_excel(file)
}
\arguments{
\item{file}{the path to the excel file we want to import the sheets from}
}
\value{
a list
}
\description{
Import all sheets from an xlsx file
}
\examples{
\dontrun{
multi_excel("/path/to/file")
}
}
