% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_csv_reader.R
\name{my_csv_reader}
\alias{my_csv_reader}
\title{Takes all the csv from a folder, import them and put them in a list}
\usage{
my_csv_reader(folder)
}
\arguments{
\item{folder}{the folder which contains all the csv files to be imported}
}
\value{
a list
}
\description{
Takes all the csv from a folder, import them and put them in a list
}
\examples{
\dontrun{
my_csv_reader(name of the folder containing the csv files)
}
}
