% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Saves data frame in csv file and outputs full path of csv}
\usage{
save_as_csv(dataset, filename, row.names = FALSE, ...)
}
\arguments{
\item{dataset}{name of the data frame to be saved as csv}

\item{filename}{name of the csv file in which the data frame will be saved}

\item{row.names}{TRUE if you want the row names of the data frame to be included in the csv file, FALSE otherwise}

\item{...}{other arguments that can be specified such as sep, na, dec...}
}
\value{
path to csv file
}
\description{
Saves data frame in csv file and outputs full path of csv
}
\examples{
\dontrun{
save_as_csv(iris, "outputiris.csv", row.names = FALSE, ...)
}
}
